quiz_web_app/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-crispy-forms==1.14.0
django-nested-admin==4.0.2
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
